Add files via upload
Allows users to input a file path, which is then assigned to the home variable. All happens in setup.

Worth noting that I've done very little in Python, and that it's only a change for main.py (Outside of dist) as I can't open the main.[nothing] file outside of terminal.